moving question creation logic to QuestionFactory object
